test: Switched from create to get_or_create in CourseEnrollmentFactory
Previously, this factory was only used in tests. So it probably never
conflicted with what is already in database.

We plan to use this factory to create data in local devstack and there
is a high chance the local devstack will already have enrollments for
(user, course) pair. Thus, what will be required is a modification.