test: V1 implementation of OEP-37
This change is a test implementation for OEP-37: Test Data. We've
written a django management command called load_test_data, that takes a
yaml file with data specifications and uses various factory_boy
factories to create that data in the database.  For MVP, the management
command only creates users and course enrollments.