squash: modified factory to use get_or_create less often
When using a factory, a user can specify any number of things.
Previously, we were using get_or_create when a subset of keys specificed
matched values in django_get_or_create list. This resulted in lots of
test failures cause when using get_or_create, factory_boy ignores all
the keys that are not in django_get_or_create. Thus those values get set
to a random value.

This change would result in get_or_create being used only when a user
specifies exactly what is in django_get_or_create list. This might not
be the best, cause this might result in data collisions in the case
where some columns are suppose to be unique.

For now, we are making this change to fix tests. Eventually, we might
have to go modify all the tests, but that will likely always be too
large of a lift for such a small detail.