Juniper upgrade - SiteConfiguration site_values - updated tests
Added test check for 'LMS_ROOT_URL'. When a SiteConfiguration record is
created, the object's `.save` method is called. Testing now checks that
'LMS_ROOT_URL' is assigned the schemd from 'settings.LMS_ROOT_URL' and
the domain of the site relating to the SiteConfiguration record is used

Also added test to make sure that SiteConfiguration.site_values
'PLATFORM_NAME' is assigned the value in
site_values.get('platform_name')