Juniper upgrade - Fix Studio local login
Studio login now performs a series of checks to attempt to authorize and
authenticate a user in the Tahoe multi-tenant environment. Given an
email, the login method checks in order of:

1. Global admin rights
2. Site admin and course access role rights

If a single user cannot be identified, authorized, and authenticated at step 1, login
fails.

If a single user cannot be identified, authorized, and authenticated at step 2, login fails.

See the docstring for `cms.djangoapps.appsembler.views.LoginView.post`