refactor!: Removing the cert_whitelist management command
[MICROBA-1052]

We are removing the cert_whitelist management command because it does
not provide a clear value with the current state of Certificate
generation on edx-platform. It adds a user to the whitelist table, but
does not generate a certificate.

We recommend using the exceptions tab in the Instructor tab of the LMS
Course.