chore: update libsass to v0.20.1
Updating libsass to latest.  Libsass is now stricter about undefined placeholder mixins, and so we had to mark a bunch of broken code as !optional.

We had upgraded libsass until it broke with this version: https://github.com/sass/libsass/releases/tag/3.3.3

That release of libsass included a fix for a missing error on undefined placeholders.

Our code was always wrong, now it just complains.  This PR fixes it.