perf: Relax load in instructor dashboard page
Relax load when bringing all the course children when fetching a course in the instructor dashboard page.

When loading the Instructor dashboard, the course is fetched with
https://github.com/edx/edx-platform/blob/3cbc5b9cdcbebb74b6f03957fcc7eb3749846add/lms/djangoapps/instructor/views/instructor_dashboard.py#L122

 Please note that the parameter **depth** is zero, so the course is fetched with no children (sections, subsections, sequentials, etc). In a subsequent part of the code, the extensions tab is loaded, specifically in :

https://github.com/edx/edx-platform/blob/3cbc5b9cdcbebb74b6f03957fcc7eb3749846add/lms/djangoapps/instructor/views/instructor_dashboard.py#L585

Which eventually call the method **get_units_with_due_date**

https://github.com/edx/edx-platform/blob/3cbc5b9cdcbebb74b6f03957fcc7eb3749846add/lms/djangoapps/instructor/views/tools.py#L121 .

This method iterates over the course and its children to find blocks with due dates. This iteration was taking too long unless the course is fetched with **depth** None, which brings all the children of the course, reason why this commit is created.