RED-2096: fix activation email server error in celery
The error happened because of the incomplete fix in https://github.com/appsembler/edx-platform/pull/884

This commit fixes the issue by getting back the mandatory
`emulate_http_request` but now with a better method to know the site
using our `appsembler.sites.utils` helpers.

This should work on celery tasks while printing the correct site domain
instead of `example.com` as it used to do before RED-1886 was fixed.