fix: don't escape assignment names in grades API
We had been pre-escaping display names like 'Math & Science' as
'Math &amp; Science" in the REST API itself - which meant that
consumers like MFEs that do their own escaping displayed the wrong
thing.

It's better for the API to just leave the string as-is. As far as
I know, this only affects the gradebook and progress pages, both
of which do their own escaping of the result from the API already.

AA-808